Move database updater to libs
